Split tests and assert inferred types with more precision
