+++ src/domain-functions.ts
@@ -1,13 +1,8 @@
 import { z } from 'https://deno.land/x/zod@v3.19.1/mod.ts'
 
-import {
-  EnvironmentError,
-  InputError,
-  InputErrors,
-  ResultError,
-} from './errors.ts'
-import { schemaError, toErrorWithMessage } from './errors.ts'
-import { isListOfSuccess, formatSchemaErrors, mergeObjects } from './utils.ts'
+import { ResultError } from './errors.ts'
+import { toErrorWithMessage } from './errors.ts'
+import { isListOfSuccess, mergeObjects } from './utils.ts'
 import type {
   DomainFunction,
   ErrorData,
@@ -21,82 +16,6 @@ import type {
 } from './types.ts'
 import type { Last } from './types.ts'
 import type { SuccessResult } from './types.ts'
-
-function makeDomainFunction<
-  Schema extends z.ZodTypeAny,
-  EnvSchema extends z.ZodTypeAny,
->(inputSchema?: Schema, environmentSchema?: EnvSchema) {
-  return function <Output>(
-    handler: (
-      input: z.infer<Schema>,
-      environment: z.infer<EnvSchema>,
-    ) => Output,
-  ) {
-    return async function (input, environment = {}) {
-      const envResult = await (
-        environmentSchema ?? z.object({})
-      ).safeParseAsync(environment)
-      const result = await (inputSchema ?? z.undefined()).safeParseAsync(input)
-
-      try {
-        if (result.success === true && envResult.success === true) {
-          return {
-            success: true,
-            data: await handler(result.data, envResult.data),
-            errors: [],
-            inputErrors: [],
-            environmentErrors: [],
-          }
-        }
-      } catch (error) {
-        if (error instanceof InputError) {
-          return {
-            success: false,
-            errors: [],
-            environmentErrors: [],
-            inputErrors: [schemaError(error.message, error.path)],
-          }
-        }
-        if (error instanceof EnvironmentError) {
-          return {
-            success: false,
-            errors: [],
-            environmentErrors: [schemaError(error.message, error.path)],
-            inputErrors: [],
-          }
-        }
-        if (error instanceof InputErrors) {
-          return {
-            success: false,
-            errors: [],
-            environmentErrors: [],
-            inputErrors: error.errors.map((e) =>
-              schemaError(e.message, e.path),
-            ),
-          }
-        }
-        if (error instanceof ResultError) return error.result
-
-        return {
-          success: false,
-          errors: [toErrorWithMessage(error)],
-          inputErrors: [],
-          environmentErrors: [],
-        }
-      }
-      return {
-        success: false,
-        errors: [],
-        inputErrors: result.success
-          ? []
-          : formatSchemaErrors(result.error.issues),
-        environmentErrors: envResult.success
-          ? []
-          : formatSchemaErrors(envResult.error.issues),
-      }
-    } as DomainFunction<Awaited<Output>>
-  }
-}
 
 function all<Fns extends DomainFunction[]>(
   ...fns: Fns
@@ -361,7 +280,6 @@ export {
   collect,
   first,
   fromSuccess,
-  makeDomainFunction,
   map,
   mapError,
   merge,
