+++ src/map-error.test.ts
@@ -1,0 +1,78 @@
+import { describe, it } from 'https://deno.land/std@0.156.0/testing/bdd.ts'
+import { assertEquals } from 'https://deno.land/std@0.160.0/testing/asserts.ts'
+import { z } from 'https://deno.land/x/zod@v3.19.1/mod.ts'
+
+import { makeDomainFunction } from './constructor.ts'
+import { mapError } from './domain-functions.ts'
+import type { DomainFunction, ErrorData } from './types.ts'
+import type { Equal, Expect } from './types.test.ts'
+
+describe('mapError', () => {
+  it('returns the result when the domain function suceeds', async () => {
+    const a = makeDomainFunction(z.object({ id: z.number() }))(
+      ({ id }) => id + 1,
+    )
+    const b = () =>
+      ({
+        errors: [{ message: 'New Error Message' }],
+        inputErrors: [{ message: 'New Input Error Message' }],
+      } as ErrorData)
+
+    const c = mapError(a, b)
+    type _R = Expect<Equal<typeof c, DomainFunction<number>>>
+
+    assertEquals(await c({ id: 1 }), {
+      success: true,
+      data: 2,
+      errors: [],
+      inputErrors: [],
+      environmentErrors: [],
+    })
+  })
+
+  it('returns a domain function function that will apply a function over the error of the first one', async () => {
+    const a = makeDomainFunction(z.object({ id: z.number() }))(
+      ({ id }) => id + 1,
+    )
+    const b = (result: ErrorData) =>
+      ({
+        errors: [{ message: 'Number of errors: ' + result.errors.length }],
+        environmentErrors: [],
+        inputErrors: [
+          {
+            message: 'Number of input errors: ' + result.inputErrors.length,
+            path: [],
+          },
+        ],
+      } as ErrorData)
+
+    const c = mapError(a, b)
+    type _R = Expect<Equal<typeof c, DomainFunction<number>>>
+
+    assertEquals(await c({ invalidInput: '1' }), {
+      success: false,
+      errors: [{ message: 'Number of errors: 0' }],
+      environmentErrors: [],
+      inputErrors: [{ message: 'Number of input errors: 1', path: [] }],
+    })
+  })
+
+  it('returns the error when the mapping function fails', async () => {
+    const a = makeDomainFunction(z.object({ id: z.number() }))(
+      ({ id }) => id + 1,
+    )
+    const b = () => {
+      throw 'failed to map'
+    }
+
+    const c = mapError(a, b)
+    type _R = Expect<Equal<typeof c, DomainFunction<number>>>
+
+    assertEquals(await c({ invalidInput: '1' }), {
+      success: false,
+      errors: [{ message: 'failed to map', exception: 'failed to map' }],
+      inputErrors: [],
+      environmentErrors: [],
+    })
+  })
+})
